Continue to add functionality to application
- Added a lot of the bare bones to the emailer. Just need to "plug"
things in.
- Some TODO list things for the Previewer. That bit needs some more
thought put into engineering it. Need more MVC
(http://stackoverflow.com/questions/5217611/the-mvc-pattern-and-swing)
- Added what I assume will be all of the functionality for the main
class, AssignmentReturner.
- The splitter now updates the roster with the file path to its
individual PDF file path. (This information should no doubt be very
useful as a Model for the MVC architecture for the AssignmentPreviewer)
- Threw the default headers into the RosterFileParser.

TODO:
- Should probably add validation for the headers in the RosterFileParser
- Add more details to the AssignmentReturner option descriptions (Like
the default headers)
- Consider adding the ability to email with just a roster file that has
the PDF_LOCATION_HEADER, along with an explanation of how to use that.
(If so, probably not something for the README.md, to avoid confusion.
Would want that in the GitHub wiki or at least just the
AssignmentEmailer JavaDoc)
- More unit Tests!
- Integration tests