Made some fix attempts, but still in a bad state.
The StudentAssignmentConfirmationController seems to be the culprit.
The indices for the roster itself and the table model are offset in
certain ways.
On one hand, if the rowOrderOfAssignments list is built according to the
table model via RosterView.getSelectedTableModelRow(), the emails are
correct to the name, the correctly named, but incorrect PDF is sent, and
the name in the email body is off on both the email and the PDF.
All this happens IF the PDFs are presented in a different order than the
roster is already in when presented in the RosterView.

When the RosterView simply used the getSelectedRow() method via the
JXTable (rosterTable variable) which is done by forcing return of
rosterTable.getSelectedRow() only at the time of PDF assignment, then
the email, name, and name used in the body of the email are correct, the
PDF is named correctly, but the PDF is INCORRECT.

Committing what I have, but switching to a git branch where I will make
a Student class to represent the rows in the roster instead of Google's
Table implementation.